@@ -249,27 +246,19 @@
     <build>
         <plugins>
             <plugin>
-                <groupId>org.codehaus.mojo</groupId>
+                <groupId>org.apache.netbeans.utilities</groupId>
                 <artifactId>nbm-maven-plugin</artifactId>
-                <version>4.1</version>
+                <version>4.8</version>
                 <extensions>true</extensions>
-                <configuration>
-                    <author>Yann D'Isanto</author>
-                    <licenseName>MIT License</licenseName>
-                    <licenseFile>MIT-LICENSE.txt</licenseFile>
-                    <keystore>private/keystore</keystore>
-                    <keystorealias></keystorealias>
-                    <keystorepassword></keystorepassword>
-                </configuration>
             </plugin>
 
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.5.1</version>
+                <version>3.12.1</version>
                 <configuration>
-                    <source>${maven.compiler.source}</source>
-                    <target>${maven.compiler.target}</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                     <compilerArgument>-Xlint:all</compilerArgument>
                 </configuration>
             </plugin>
